+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### New
 - Added Unbound data. [#20](https://github.com/69MichelleDB/ToNCodes/issues/20)
+- Added cryptography and requests to About window.
+- Added debug window, to enable it change the debug-window node in the config.xml file to 1. [#26](https://github.com/69MichelleDB/ToNCodes/issues/26)
 
 ### Fixes
 - In case the `control file` gets corrupted or lost, it'll be deleted and a fresh one will be created. Next, ToNCodes will read all available Log files to first find and insert any missing codes and second to regenerate the `control file` with the missing cursor and dates data. [#21](https://github.com/69MichelleDB/ToNCodes/issues/21)